commit code 4/3

+++ AnService_Capstone.Core/Entities/TblContract.cs
@@ -14,8 +14,8 @@ namespace AnService_Capstone.Core.Entities
         public string ContractUrl { get; set; }
         public DateTime? ContractStartDate { get; set; }
         public DateTime? ContractEndDate { get; set; }
-        public double? ContractDeposit { get; set; }
-        public double? ContractTotalPrice { get; set; }
+        public decimal? ContractDeposit { get; set; }
+        public decimal? ContractTotalPrice { get; set; }
         public int ContractStatus { get; set; }
         public DateTime? ContractCreateDate { get; set; }
         public DateTime? ContractUpdateDate { get; set; }
